Uploading v1.1.0
git-svn-id: https://plugins.svn.wordpress.org/stripe/trunk@902236 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ class-stripe-checkout.php
@@ -21,7 +21,7 @@ class Stripe_Checkout {
 	 *
 	 * @var     string
 	 */
-	protected $version = '1.0.1';
+	protected $version = '1.1.0';
 
 	/**
 	 * Unique identifier for your plugin.
@@ -278,6 +278,9 @@ class Stripe_Checkout {
 		
 		// Include shortcode functions
 		include_once( 'includes/shortcodes.php' );
+		
+		// Hooks examples
+		//include_once( 'includes/hooks-examples.php' );
 	}
 
 	/**
@@ -288,7 +291,7 @@ class Stripe_Checkout {
 	 * @return    string
 	 */
 	public static function get_plugin_title() {
-		return __( 'Simple Stripe Checkout', 'sc' );
+		return __( 'Stripe Checkout', 'sc' );
 	}
 
 
@@ -338,11 +341,16 @@ class Stripe_Checkout {
 		if ( false == get_option( 'sc_show_admin_install_notice' ) )
 			return;
 
-		// At this point show install notice.
-		include_once( 'views/admin-install-notice.php' );
+		// Delete stored value if "hide" button click detected (custom querystring value set to 1).
+		// or if on a PIB admin page. Then exit.
+		if ( ! empty( $_REQUEST['sc-dismiss-install-nag'] ) || $this->viewing_this_plugin() ) {
+			delete_option( 'sc_show_admin_install_notice' );
+			return;
+		}
 
-		// Delete stored value to hide since we only want them to view it once.
-		delete_option( 'sc_show_admin_install_notice' );
-		return;
+		// At this point show install notice. Show it only on the plugin screen.
+		if( get_current_screen()->id == 'plugins' ) {
+			include_once( 'views/admin-install-notice.php' );
+		}
 	}
 }
